refactor: 스웨거 패키지를 컨트롤러 패키지 하위로 이동
