+++ backend/src/main/java/mouda/backend/auth/presentation/controller/AuthController.java
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
+import mouda.backend.auth.business.AuthService;
+import mouda.backend.auth.presentation.controller.swagger.AuthSwagger;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
-import mouda.backend.auth.presentation.swagger.AuthSwagger;
-import mouda.backend.auth.business.AuthService;
 import mouda.backend.common.response.RestResponse;
 
 @RestController
