+++ backend/src/main/java/mouda/backend/darakbang/implement/InvitationCodeGenerator.java
@@ -1,4 +1,4 @@
-package mouda.backend.darakbang.business;
+package mouda.backend.darakbang.implement;
 
 import java.security.SecureRandom;
