+++ backend/src/main/java/mouda/backend/please/presentation/controller/swagger/InterestSwagger.java
@@ -1,4 +1,4 @@
-package mouda.backend.please.presentation.swagger;
+package mouda.backend.please.presentation.controller.swagger;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
